OrderController.php, OrderItem.php

+++ onlineshop-app/app/Http/Controllers/OrderController.php
@@ -3,16 +3,18 @@
 namespace App\Http\Controllers;
 
 use App\Models\Order;
-use Illuminate\Http\Request;
 use App\Http\Resources\OrderResource;
 use App\Http\Requests\OrderRequest;
 use Illuminate\Http\Response;
+use App\Models\OrderItem;
+use Illuminate\Support\Facades\DB;
 
 class OrderController extends Controller
 {
-    /*public function __construct(){
-        $this->authorizeResource(Order::class);
-    }*/
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index']);
+    }
 
     /**
      * Display a listing of the resource.
@@ -32,7 +34,20 @@ class OrderController extends Controller
      */
     public function store(OrderRequest $request)
     {
-        $create_order = Order::create($request->validated());
+        $request->validated();
+
+        $create_order = Order::create([
+            'user_id' => auth()->user()->id,
+            'status' => $request->status,
+            'comment' => $request->comment,
+            'address' => $request->address,
+        ]);
+
+        DB::table('product_order')->insert([
+            'product_id' => $request->product_id,
+            'order_id' => $create_order->id,
+        ]);
+
         return new OrderResource($create_order);
     }
 
@@ -42,9 +57,12 @@ class OrderController extends Controller
      * @param int $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
+    public function show(Order $order)
     {
-        return new OrderResource(Order::findOrFail($id));
+        if (auth()->user()->id === $order->user_id || auth()->user()->isAdmin()) {
+            return new OrderResource(Order::findOrFail($order->id));
+        }
+        return response()->json(['message' => 'Action Forbidden']);
     }
 
     /**
@@ -56,8 +74,23 @@ class OrderController extends Controller
      */
     public function update(OrderRequest $request, Order $order)
     {
-        $order->update($request->validated());
-        return new OrderResource($order);
+        if (auth()->user()->id === $order->user_id) {
+
+            $request->validated();
+            $order->update([
+                'status' => $request->status,
+                'comment' => $request->comment,
+                'address' => $request->address,
+            ]);
+            DB::table('product_order')
+                ->where('order_id', $order->id)
+                ->update([
+                    'product_id' => $request->product_id,
+                ]);
+            return new OrderResource($order);
+        }
+
+        return response()->json(['message' => 'Action Forbidden']);
     }
 
     /**
@@ -68,7 +101,31 @@ class OrderController extends Controller
      */
     public function destroy(Order $order)
     {
-        $order->delete();
-        return response(null, Response::HTTP_NO_CONTENT);
+        if (auth()->user()->id === $order->user_id || auth()->user()->isAdmin()) {
+            DB::table('product_order')->where('order_id', "=", $order->id)->delete();
+            DB::table('order_items')->where('order_id', "=", $order->id)->delete();
+            $order->delete();
+            return response(null, Response::HTTP_NO_CONTENT);
+        }
+        return response()->json(['message' => 'Action Forbidden']);
+    }
+
+    public function getOrderItem()
+    {
+        $orderId = DB::table('orders')->where('user_id', auth()->user()->id)->pluck('id');
+        $products_id = DB::table('product_order')->where('order_id', $orderId[0])->pluck('product_id');
+        $quantity = 0;
+        $price = 0;
+        for (; $quantity < count($products_id); $quantity++) {
+            $price += DB::table('products')
+                ->where('id', $products_id[$quantity])
+                ->sum('price');
+        }
+        return OrderItem::create([
+            'order_id' => $orderId[0],
+            'quantity' => $quantity,
+            'price' => $price,
+        ]);
+
     }
 }
